Isolate modem preset CLI tests and reset state
